# next-frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos": {
      "post": {
        "description": "Pre-registers the video as a draft on the caller's channel and opens a multipart upload, returning the plan the client follows to send parts straight to object storage.",
        "operationId": "VideosController_initiateUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload initiated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "upload_id": {
                      "type": "string"
                    },
                    "part_size_bytes": {
                      "type": "integer"
                    },
                    "part_count": {
                      "type": "integer"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "File exceeds the maximum upload size",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Content type is not a video type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Initiate a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/upload/parts": {
      "post": {
        "description": "Issues presigned UploadPart URLs for the requested part numbers. This is also the resume path: ask only for the parts still missing.",
        "operationId": "VideosController_presignParts",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignPartsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned part URLs",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "parts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "part_number": {
                            "type": "integer"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this public id belongs to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video has no active multipart upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Presign upload parts",
        "tags": [
          "videos"
        ]
      },
      "get": {
        "description": "Returns the parts the object storage already holds for the active multipart upload, with their ETags. This is the resume path after a page reload: the client signs only the missing parts and completes with server-sourced ETags.",
        "operationId": "VideosController_listUploadedParts",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parts already received by the storage",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "parts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "part_number": {
                            "type": "integer"
                          },
                          "etag": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this public id belongs to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video has no active multipart upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the parts already uploaded",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/upload/complete": {
      "post": {
        "description": "Finalizes the multipart upload and, in the same transaction, moves the video to `processing` and enqueues the processing job.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload completed and processing enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "processing"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this public id belongs to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video has no active multipart upload (`INVALID_UPLOAD_STATE`), or the stored parts do not add up to the declared `size_bytes` (`UPLOAD_SIZE_MISMATCH`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Complete a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/upload": {
      "delete": {
        "description": "Aborts the multipart upload and discards the draft, releasing the storage it held.",
        "operationId": "VideosController_abortUpload",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Upload aborted and draft discarded"
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this public id belongs to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video has no active multipart upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Abort a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/status": {
      "get": {
        "description": "Owner-only view of the `draft → processing → ready | failed` lifecycle. Metadata and a presigned thumbnail URL appear once the video is ready; the failure reason appears when it failed.",
        "operationId": "VideosController_getStatus",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current processing status",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "processing",
                        "ready",
                        "failed"
                      ]
                    },
                    "duration_seconds": {
                      "type": "integer"
                    },
                    "width": {
                      "type": "integer"
                    },
                    "height": {
                      "type": "integer"
                    },
                    "thumbnail_url": {
                      "type": "string"
                    },
                    "failure_reason": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this public id belongs to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Poll the processing status",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}": {
      "get": {
        "description": "Public metadata for a ready video — the API behind the watch page. A video that is not ready is invisible to anyone but its owner.",
        "operationId": "VideosController_getPublicVideo",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public video metadata",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "example": "ready"
                    },
                    "duration_seconds": {
                      "type": "integer",
                      "nullable": true
                    },
                    "width": {
                      "type": "integer",
                      "nullable": true
                    },
                    "height": {
                      "type": "integer",
                      "nullable": true
                    },
                    "thumbnail_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown public id, or not ready and the caller is not the owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Resolve public video metadata",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream": {
      "get": {
        "description": "Authorizes playback and returns a short-lived presigned URL the player reads directly from object storage, with native HTTP Range support for seeking.",
        "operationId": "VideosController_getStreamUrl",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public id of the video",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Range-capable presigned playback URL",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown public id, or not ready and the caller is not the owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The owner requested playback of a video that is not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Issue a playback URL",
        "tags": [
          "videos"
        ]
      }
    },
...
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InitiateUploadDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Original file name — only used to derive the storage key",
            "example": "clip.mp4"
          },
          "content_type": {
            "type": "string",
            "description": "MIME type of the upload; must be a `video/*` type",
            "example": "video/mp4"
          },
          "size_bytes": {
            "type": "number",
            "description": "Total size of the file in bytes (max 10 GiB)",
            "example": 52428800,
            "minimum": 1
          },
          "part_size_bytes": {
            "type": "number",
            "description": "Desired part size in bytes; at least 5242880 (5 MiB), the S3 minimum for every part but the last",
            "example": 67108864,
            "minimum": 5242880
          }
        },
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ]
      },
      "PresignPartsDto": {
        "type": "object",
        "properties": {
          "part_numbers": {
            "description": "Part numbers to presign. On resume, send only the parts still missing.",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "part_numbers"
        ]
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}